+++ lib/rpmfi.h
@@ -143,8 +143,8 @@ struct rpmfi_s {
 
     int astriplen;
     int striplen;
-    unsigned int archivePos;
-    unsigned int archiveSize;
+    unsigned long long archivePos;
+    unsigned long long archiveSize;
     mode_t dperms;		/*!< Directory perms (0755) if not mapped. */
     mode_t fperms;		/*!< File perms (0644) if not mapped. */
 /*@only@*/ /*@null@*/
